new demo

+++ app/src/main/res/raw/skybox_fragment.glsl
@@ -1,5 +1,5 @@
-#version 300 es
-precision mediump float;
+#version 320 es
+precision highp float;
 
 out vec4 FragColor;
